Add more tests. Change unset error msg. Change error code for non-existent file
