+++ builtins/exit.c
@@ -3,27 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   exit.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lalex-ku <lalex-ku@42sp.org.br>            +#+  +:+       +#+        */
+/*   By: sguilher <sguilher@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/12 14:02:34 by lalex-ku          #+#    #+#             */
-/*   Updated: 2022/05/31 16:18:55 by lalex-ku         ###   ########.fr       */
+/*   Updated: 2022/05/31 17:02:57 by sguilher         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-#define MAX_LONG 9223372036854775807
-#define MIN_LONG -9223372036854775807
-
-static void	print_error_and_exit(char *command, char *msg, int error)
-{
-	ft_putstr_fd("minishell: ", STDERR_FILENO); // TODO: Colocar na ft_printf
-	ft_putstr_fd(command, STDERR_FILENO);
-	ft_putstr_fd(": ", STDERR_FILENO);
-	ft_putstr_fd(msg, STDERR_FILENO);
-	ft_putstr_fd("\n", STDERR_FILENO);
-	exit(error);
-}
 
 static int	str_is_long(const char *str)
 {
@@ -51,14 +38,14 @@ int	builtin_exit(char **args, t_env **minienv)
 {
 	char	*exit_status;
 	long	number;
-	
+
 	rl_clear_history();
 	free_minienv(minienv);
 	ft_putstr_fd("exit\n", STDOUT_FILENO);
 	if (!args[0] || !args[1])
 	{
 		free_array(args);
-		exit(EXIT_SUCCESS);	
+		exit(EXIT_SUCCESS);
 	}
 	exit_status = args[1];
 	if (!str_is_long(exit_status))
