+++ tests/manual_tests
@@ -3,20 +3,35 @@ echo -n hello ✅
 
 CTRL+C
 CTRL+D
-CTRL+\ 
+CTRL+\
 CTRL+C, CTRL+C, CTRL+C, exit ❌
 
 env
 env hello world # -> We need to implement "env with no options or arguments", the output here is optional
 
-export
-export hello
 export HELLO=123
-unset
-unset HELLO 
-unset HELLO=
-cd # Not mandatory, we need to imlement only "cd with only a relative or absolute path"
-cd obj, echo $PWD $OLDPWD 
+export HELLO="beautiful world"
 
-exit 9223372036854775807
-exit 9223372036854775808
+cd # Not mandatory, we need to imlement only "cd with only a relative or absolute path"
+cd obj, echo $PWD $OLDPWD
+
+env | grep USER || grep lucas # não tem que implementar o ||
+# verificar check de um caminho que é um link simbólico (por exemplo, um fd de um processo) - se a stat funciona ou precisaria usar a lstat
+# executar ./loop.c e verificar o exit code
+#cat <<doc1 | tr a b | <<doc2 ls
+
+# exit status quando roda um comando inválido e logo em seguida roda um espaço
+oi
+
+echo $?
+
+#exit status quando roda um comando invalido e logo em seguida expande uma variável vazia
+asd; $HELLO; echo $?
+
+#Testar se o worskpaces tem OLD_PWD setado
+# mv, cp, rm
+
+cat minishell.h | grep $");" # não precisa ser igual ao do bash porque só temos que lidar com o $ como expansão de variável
+
+#extra: variáveis locais
+aa==vv
