+++ tests/tests_list
@@ -18,10 +18,16 @@ $PWD
 pwd
 pwd oi
 export hello
+export HELLO=123 | env | grep HELLO
 export HELLO=123
+export HELLO=123 kjkjkllkç
+export HELLO="123 kjkjkllkç"
+export HELLO.=123
+export HELLO-=123
 unset
-unset HELLO
 unset HELLO=
+unset HELLO
+unset asd=123 | grep oi
 cd $PWD
 cd 123123
 exit 123
@@ -29,6 +35,20 @@ exit 298
 exit -1000
 exit hello
 exit 42 world
+exit 9223372036854775807
+exit 9223372036854775808
+exit -9223372036854775807
+exit -9223372036854775808
+exit -9223372036854775809
 ./tests/test_files/invalid_permission
 invalid_command
-export HELLO=123 | env | grep HELLO
+./b.out
+b.out
+"aa"
+./obj
+minishell.h
+oi
+ls
+cat ./tests/test_files/infile_big
+cat ./tests/test_files/infile_big | grep oi
+cat minishell.h | grep ");"$
